Reorganised the optimiser files into a new filter section. Moved around the utils so that the common stuff is in top level utils and generated set utils has stuff specific to generated sets. Deleted some stuff that was no longer used anywhere. Fixed an issue where stats with perk mixes wasn't correct for armour 1.0 stats.
