+++ src/app/loadout-builder/filter/LockArmorAndPerks.m.scss.d.ts
