+++ src/app/loadout-builder/filter/LockedArmor2ModIcon.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
-import ClosableContainer from './ClosableContainer';
+import ClosableContainer from '../ClosableContainer';
 import { SocketDetailsMod } from 'app/item-popup/SocketDetails';
-import { LockedArmor2Mod, LockedModBase } from './types';
+import { LockedArmor2Mod, LockedModBase } from '../types';
 import styles from './LockedItem.m.scss';
 import { D2ManifestDefinitions } from 'app/destiny2/d2-definitions';
