+++ src/app/loadout-builder/filter/LockedItem.m.scss
