+++ src/app/loadout-builder/filter/ModPickerFooter.m.scss
@@ -1,12 +1,12 @@
-@import '../variables.scss';
+@import '../../variables.scss';
 
 .footer {
-  composes: flexRow from '../dim-ui/common.m.scss';
+  composes: flexRow from '../../dim-ui/common.m.scss';
   align-items: center;
 }
 
 .selectedMods {
-  composes: flexWrap from '../dim-ui/common.m.scss';
+  composes: flexWrap from '../../dim-ui/common.m.scss';
   flex: 1;
   align-items: center;
