+++ src/app/loadout-builder/filter/ModPickerHeader.m.scss
@@ -1,4 +1,4 @@
-@import '../variables.scss';
+@import '../../variables.scss';
 
 .tabs {
   composes: move-popup-tabs from global;
