+++ src/app/loadout-builder/filter/PerksForBucket.tsx
@@ -1,15 +1,15 @@
 import React from 'react';
 import { DestinyInventoryItemDefinition } from 'bungie-api-ts/destiny2';
 import { InventoryBucket } from 'app/inventory/inventory-buckets';
-import { LockedItemType, BurnItem } from './types';
+import { LockedItemType, BurnItem } from '../types';
 import {
   SelectableBurn,
   SelectablePerk,
   SelectableMod,
-} from './locked-armor/SelectableBungieImage';
+} from '../locked-armor/SelectableBungieImage';
 import styles from './PerksForBucket.m.scss';
 import { DimItem } from 'app/inventory/item-types';
-import { getFilteredPerksAndPlugSets } from './generated-sets/utils';
+import { getFilteredPerksAndPlugSets } from '../utils';
 import { D2ManifestDefinitions } from 'app/destiny2/d2-definitions';
 
 /**
