+++ src/app/loadout-builder/generated-sets/GeneratedSet.tsx
@@ -5,7 +5,8 @@ import GeneratedSetButtons from './GeneratedSetButtons';
 import GeneratedSetItem from './GeneratedSetItem';
 import { powerIndicatorIcon, AppIcon, faExclamationTriangle } from '../../shell/icons';
 import _ from 'lodash';
-import { getNumValidSets, calculateTotalTier, statTier, sumEnabledStats } from './utils';
+import { getNumValidSets, calculateTotalTier, sumEnabledStats } from './utils';
+import { statTier } from '../utils';
 import { D2ManifestDefinitions } from 'app/destiny2/d2-definitions';
 import BungieImage from 'app/dim-ui/BungieImage';
 import { DestinyStatDefinition } from 'bungie-api-ts/destiny2';
@@ -14,7 +15,7 @@ import { t } from 'app/i18next-t';
 import styles from './GeneratedSet.m.scss';
 import { editLoadout } from 'app/loadout/LoadoutDrawer';
 import { Loadout } from 'app/loadout/loadout-types';
-import { assignModsToArmorSet } from './mod-utils';
+import { assignModsToArmorSet } from '../mod-utils';
 import { Armor2ModPlugCategories } from 'app/utils/item-utils';
 import { LoadoutBuilderAction } from '../loadoutBuilderReducer';
 import PressTip from 'app/dim-ui/PressTip';
