+++ src/app/loadout-builder/mod-utils.ts
@@ -1,6 +1,6 @@
-import { DimItem } from './../../inventory/item-types';
+import { DimItem } from '../inventory/item-types';
 import _ from 'lodash';
-import { LockableBuckets, LockedArmor2ModMap, LockedArmor2Mod } from './../types';
+import { LockableBuckets, LockedArmor2ModMap, LockedArmor2Mod } from './types';
 import { DestinyEnergyType } from 'bungie-api-ts/destiny2';
 import { getSpecialtySocketMetadata, Armor2ModPlugCategories } from 'app/utils/item-utils';
