+++ src/app/loadout-builder/utils.ts
@@ -1,8 +1,293 @@
 import _ from 'lodash';
-import { DimPlug, DimItem } from 'app/inventory/item-types';
+import { DimPlug, DimItem, D2Item, DimSocket } from 'app/inventory/item-types';
 import { getMasterworkSocketHashes } from 'app/utils/socket-utils';
 import { DestinySocketCategoryStyle } from 'bungie-api-ts/destiny2';
-import { statValues, LockedItemType, LockedArmor2Mod } from './types';
+import { statValues, LockedItemType, LockedMod, LockedArmor2Mod } from './types';
+import {
+  DestinyInventoryItemDefinition,
+  TierType,
+  DestinyItemSubType,
+  DestinyEnergyType,
+} from 'bungie-api-ts/destiny2';
+import { getSpecialtySocketMetadata } from 'app/utils/item-utils';
+
+/**
+ * Plug item hashes that should be excluded from the list of selectable perks.
+ */
+const unwantedSockets = new Set([
+  3313201758, // Mobility, Restorative, and Resilience perks
+  1514141499, // Void damage resistance
+  1514141501, // Arc damage resistance
+  1514141500, // Solar damage resistance
+  2973005342, // Shaders
+  3356843615, // Ornaments
+  2457930460, // Empty masterwork slot
+]);
+const unwantedCategories = new Set([
+  1742617626, // ItemCategory "Armor Mods: Ornaments"
+  1875601085, // ItemCategory "Armor Mods: Glow Effects"
+  1404791674, // ItemCategory "Ghost Mods: Projections"
+]);
+
+/**
+ *  Filter out plugs that we don't want to show in the perk picker.
+ */
+export function filterPlugs(socket: DimSocket) {
+  if (!socket.plug) {
+    return false;
+  }
+
+  const plugItem = socket.plug.plugItem;
+  if (!plugItem || !plugItem.plug) {
+    return false;
+  }
+
+  // Armor 2.0 mods
+  if (socket.plug.plugItem.collectibleHash) {
+    return false;
+  }
+
+  if (
+    plugItem.itemSubType === DestinyItemSubType.Ornament ||
+    plugItem.itemSubType === DestinyItemSubType.Shader
+  ) {
+    return false;
+  }
+
+  // Remove unwanted sockets by category hash
+  if (
+    unwantedSockets.has(plugItem.plug.plugCategoryHash) ||
+    plugItem.itemCategoryHashes?.some((h) => unwantedCategories.has(h))
+  ) {
+    return false;
+  }
+
+  // Remove Archetype/Inherit perk
+  if (
+    plugItem.plug.plugCategoryHash === 1744546145 &&
+    plugItem.inventory.tierType !== 6 // keep exotics
+  ) {
+    return false;
+  }
+
+  // Remove empty mod slots
+  if (
+    plugItem.plug.plugCategoryHash === 3347429529 &&
+    plugItem.inventory.tierType === TierType.Basic
+  ) {
+    return false;
+  }
+
+  // Remove masterwork mods and energy mods
+  if (plugItem.plug.plugCategoryIdentifier.match(/masterworks/)) {
+    return false;
+  }
+
+  // Remove empty sockets, which are common tier
+  if (plugItem.inventory.tierType === TierType.Common) {
+    return false;
+  }
+
+  // Only real mods
+  if (
+    !socket.isPerk &&
+    (plugItem.inventory.bucketTypeHash !== 3313201758 || !plugItem.inventory.recoveryBucketTypeHash)
+  ) {
+    return false;
+  }
+
+  return true;
+}
+
+/**
+ * Add a locked item to the locked item list for a bucket.
+ */
+export function addLockedItem(
+  lockedItem: LockedItemType,
+  locked: readonly LockedItemType[] = []
+): readonly LockedItemType[] | undefined {
+  // Locking an item clears out the other locked properties in that bucket
+  if (lockedItem.type === 'item') {
+    return [lockedItem];
+  }
+
+  // there can only be one burn type per bucket
+  if (lockedItem.type === 'burn') {
+    const newLockedItems = locked.filter((li) => li.type !== 'burn');
+    newLockedItems.push(lockedItem);
+    return newLockedItems;
+  }
+
+  // Only add if it's not already there.
+  if (!locked.some((existing) => lockedItemsEqual(existing, lockedItem))) {
+    const newLockedItems = Array.from(locked);
+    newLockedItems.push(lockedItem);
+    return newLockedItems;
+  }
+
+  return locked.length === 0 ? undefined : locked;
+}
+
+/**
+ * Remove a locked item from the locked item list for a bucket.
+ */
+export function removeLockedItem(
+  lockedItem: LockedItemType,
+  locked: readonly LockedItemType[] = []
+): readonly LockedItemType[] | undefined {
+  // Filter anything equal to the passed in item
+  const newLockedItems = locked.filter((existing) => !lockedItemsEqual(existing, lockedItem));
+  return newLockedItems.length === 0 ? undefined : newLockedItems;
+}
+
+export function lockedItemsEqual(first: LockedItemType, second: LockedItemType) {
+  switch (first.type) {
+    case 'item':
+      return second.type === 'item' && first.item.id === second.item.id;
+    case 'exclude':
+      return second.type === 'exclude' && first.item.id === second.item.id;
+    case 'mod':
+      return second.type === 'mod' && first.mod.hash === second.mod.hash;
+    case 'perk':
+      return second.type === 'perk' && first.perk.hash === second.perk.hash;
+    case 'burn':
+      return second.type === 'burn' && first.burn.dmg === second.burn.dmg;
+  }
+}
+
+/**
+ * filteredPerks:
+ * The input perks, filtered down to perks on items that also include the other selected perks in that bucket.
+ * For example, if you'd selected "heavy ammo finder" for class items it would only include perks that are on
+ * class items that also had "heavy ammo finder".
+ *
+ * filteredPlugSetHashes:
+ * Plug set hashes that contain the mods that can slot into items that can also slot the other selected mods in that bucket.
+ * For example, if you'd selected "scout rifle loader" for gauntlets it would only include mods that can slot on
+ * gauntlets that can also slot "scout rifle loader".
+ */
+export function getFilteredPerksAndPlugSets(
+  locked: readonly LockedItemType[] | undefined,
+  items: readonly DimItem[]
+) {
+  const filteredPlugSetHashes = new Set<number>();
+  const filteredPerks = new Set<DestinyInventoryItemDefinition>();
+
+  if (!locked) {
+    return {};
+  }
+
+  for (const item of items) {
+    // flat list of plugs per item
+    const itemPlugs: DestinyInventoryItemDefinition[] = [];
+    // flat list of plugSetHashes per item
+    const itemPlugSets: number[] = [];
+
+    if (item.isDestiny2() && item.sockets) {
+      for (const socket of item.sockets.sockets) {
+        // Populate mods
+        if (!socket.isPerk) {
+          if (socket.socketDefinition.reusablePlugSetHash) {
+            itemPlugSets.push(socket.socketDefinition.reusablePlugSetHash);
+          } else if (socket.socketDefinition.randomizedPlugSetHash) {
+            itemPlugSets.push(socket.socketDefinition.randomizedPlugSetHash);
+          }
+        }
+
+        // Populate plugs
+        if (filterPlugs(socket)) {
+          socket.plugOptions.forEach((option) => {
+            itemPlugs.push(option.plugItem);
+          });
+        }
+      }
+    }
+
+    // The item must be able to slot all mods
+    let matches = true;
+    for (const lockedItem of locked) {
+      if (lockedItem.type === 'mod') {
+        const mod = lockedItem.mod;
+        if (item.isDestiny2() && matchesEnergy(item, mod)) {
+          const plugSetIndex = itemPlugSets.indexOf(lockedItem.plugSetHash);
+          if (plugSetIndex >= 0) {
+            // Remove this plugSetHash from the list because it is now "occupied"
+            itemPlugSets.splice(plugSetIndex, 1);
+          } else {
+            matches = false;
+            break;
+          }
+        } else {
+          matches = false;
+          break;
+        }
+      }
+    }
+
+    // for each item, look to see if all perks match locked
+    matches =
+      matches &&
+      locked.every(
+        (locked) =>
+          locked.type !== 'perk' || itemPlugs.some((plug) => plug.hash === locked.perk.hash)
+      );
+
+    // It matches all perks and plugs
+    if (matches) {
+      for (const plugSetHash of itemPlugSets) {
+        filteredPlugSetHashes.add(plugSetHash);
+      }
+      for (const itemPlug of itemPlugs) {
+        filteredPerks.add(itemPlug);
+      }
+    }
+  }
+
+  return { filteredPlugSetHashes, filteredPerks };
+}
+
+function matchesEnergy(item: D2Item, mod: DestinyInventoryItemDefinition) {
+  return (
+    !mod.plug ||
+    !mod.plug.energyCost ||
+    !item.energy ||
+    mod.plug.energyCost.energyType === item.energy.energyType ||
+    mod.plug.energyCost.energyType === DestinyEnergyType.Any
+  );
+}
+
+/**
+ * Can this mod be slotted onto this item?
+ */
+export function canSlotMod(item: DimItem, lockedItem: LockedMod) {
+  const mod = lockedItem.mod;
+  return (
+    item.isDestiny2() &&
+    matchesEnergy(item, mod) &&
+    // is a seasonal mod and item has correct socket
+    (getSpecialtySocketMetadata(item)?.compatiblePlugCategoryHashes.includes(
+      lockedItem.mod.plug.plugCategoryHash
+    ) ||
+      // or matches socket plugsets
+      item.sockets?.sockets.some(
+        (socket) =>
+          (socket.socketDefinition.reusablePlugSetHash &&
+            lockedItem.plugSetHash === socket.socketDefinition.reusablePlugSetHash) ||
+          (socket.socketDefinition.randomizedPlugSetHash &&
+            lockedItem.plugSetHash === socket.socketDefinition.randomizedPlugSetHash)
+      ))
+  );
+}
+
+/** Whether this item is eligible for being in loadout builder */
+export function isLoadoutBuilderItem(item: DimItem) {
+  // Armor and Ghosts
+  return item.bucket.inArmor;
+}
+
+export function statTier(stat: number) {
+  return Math.floor(stat / 10);
+}
 
 /**
  * This is a wrapper for the awkward helper used by
@@ -33,7 +318,7 @@ export function generateMixesFromPerks(
  * returns all the mixes. This is like this so we can share this complicated bit of logic
  * and not get it out of sync.
  */
-export function generateMixesFromPerksOrStats(
+function generateMixesFromPerksOrStats(
   item: DimItem,
   assumeArmor2IsMasterwork: boolean | null,
   lockedModStats: { [statHash: number]: number },
@@ -46,7 +331,9 @@ export function generateMixesFromPerksOrStats(
     return [];
   }
 
-  const mixes: number[][] = [getBaseStatValues(item, assumeArmor2IsMasterwork, lockedModStats)];
+  const mixes: number[][] = [
+    getStatValuesWithModsAndMW(item, assumeArmor2IsMasterwork, lockedModStats),
+  ];
 
   const altPerks: (DimPlug[] | null)[] = [null];
 
@@ -84,7 +371,11 @@ export function generateMixesFromPerksOrStats(
   return mixes;
 }
 
-export function getBaseStatValues(
+/**
+ * Gets the stat values of an item with masterwork and locked mods considered.
+ * Unfortunately this needs to be kept in line with getStatValuesWithModsAndMWProcess in ./processWorker/process.ts.
+ */
+export function getStatValuesWithModsAndMW(
   item: DimItem,
   assumeMasterwork: boolean | null,
   lockedModStats?: { [statHash: number]: number }
@@ -93,7 +384,7 @@ export function getBaseStatValues(
   const baseStats = {};
 
   for (const statHash of statValues) {
-    baseStats[statHash] = stats[statHash]?.value || 0;
+    baseStats[statHash] = stats[statHash]?.base || 0;
   }
 
   // Checking energy tells us if it is Armour 2.0
@@ -111,10 +402,10 @@ export function getBaseStatValues(
     for (const socket of item.sockets.sockets) {
       const plugHash = socket?.plug?.plugItem?.hash ?? NaN;
 
-      if (socket.plug?.stats && !masterworkSocketHashes.includes(plugHash)) {
+      if (socket.plug?.stats && masterworkSocketHashes.includes(plugHash)) {
         for (const statHash of statValues) {
           if (socket.plug.stats[statHash]) {
-            baseStats[statHash] -= socket.plug.stats[statHash];
+            baseStats[statHash] += socket.plug.stats[statHash];
           }
         }
       }
